+++ src/ai/backend/manager/models/keypair.py
@@ -12,6 +12,7 @@ from typing import (
     Tuple,
     TypedDict,
 )
+import uuid
 
 from aiopg.sa.connection import SAConnection
 from aiopg.sa.result import RowProxy
@@ -46,6 +47,7 @@ from .user import UserRole
 __all__: Sequence[str] = (
     'keypairs',
     'KeyPair', 'KeyPairList',
+    'UserInfo',
     'KeyPairInput',
     'CreateKeyPair', 'ModifyKeyPair', 'DeleteKeyPair',
     'Dotfile', 'MAXIMUM_DOTFILE_SIZE',
@@ -88,6 +90,39 @@ keypairs = sa.Table(
 )
 
 
+class UserInfo(graphene.ObjectType):
+    email = graphene.String()
+    full_name = graphene.String()
+
+    @classmethod
+    def from_row(
+        cls,
+        context: Mapping[str, Any],
+        row: RowProxy,
+    ) -> Optional[UserInfo]:
+        if row is None:
+            return None
+        return cls(email=row['email'], full_name=row['full_name'])
+
+    @classmethod
+    async def batch_load_by_uuid(
+        cls,
+        context: Mapping[str, Any],
+        user_uuids: List[uuid.UUID],
+    ) -> Sequence[Optional[UserInfo]]:
+        async with context['dbpool'].acquire() as conn:
+            from .user import users
+            query = (
+                sa.select([users.c.uuid, users.c.email, users.c.full_name])
+                .select_from(users)
+                .where(users.c.uuid.in_(user_uuids))
+            )
+            return await batch_result(
+                context, conn, query, cls,
+                user_uuids, lambda row: row['uuid'],
+            )
+
+
 class KeyPair(graphene.ObjectType):
 
     class Meta:
@@ -114,10 +149,20 @@ class KeyPair(graphene.ObjectType):
         status=graphene.String(),
     )
 
+    user_info = graphene.Field(lambda: UserInfo)
+
     # Deprecated
     concurrency_limit = graphene.Int(
         deprecation_reason='Moved to KeyPairResourcePolicy object as '
                            'max_concurrent_sessions field.')
+
+    async def resolve_user_info(
+        self,
+        info: graphene.ResolveInfo,
+    ) -> UserInfo:
+        manager = info.context['dlmgr']
+        loader = manager.get_loader('UserInfo.by_uuid')
+        return await loader.load(self.user)
 
     @classmethod
     def from_row(